modify: proc.Call() return uintptr to adapt pointer

+++ proc/proc.go
@@ -2,7 +2,7 @@ package proc
 
 type Proc uintptr
 
-func (p Proc) Call(args ...uintptr) int {
+func (p Proc) Call(args ...uintptr) uintptr {
 	return Call(uintptr(p), args...)
 }
